skillgui.cpp: revise command history queue ordering
When executing commands in your command history, the commands are
not moved to the top of the history list and can therefore be
removed from the list before older commands. This revision deletes
all duplicates from the command history list and then prepends the
latest command at the top of the list.