SkillGUI: hide tabs and toolbar if only single entry
This takes the ide of Kyle's history_update branch and implements it in
a way compatible with Fawkes SkillGUI. In ROS mode the toolbar is simply
hiden. The tab headers are hidden if there is only single page, which is
the case for ROS but not for Fawkes (where we also have log view and
plugin pages).