SkillGUI: tune and smooth follow mode
The follow motion is a little faster now and smoother, it does not have
the "click-in at end" effect anymore. Fix some little glitches, for
example do not interfere with mouse motions. Have maximum dt, float
casting for proper time calculations, on startup if speed is 0.0 set dt
and speed explicitly once (dt is not meaningful at that point
otherwise).